Move IndexingProgress DTO from _mcp to _indexing
- Relocate IndexingProgress dataclass to _indexing/progress.py.
- Update progress reporter and worker orchestrator imports.
- _mcp.state_manager imports from the new canonical location.

make format-check, make lint, make type-check pass.
Full test suite: 1432 passed, 17 skipped, 4 xpassed.
